Move user mention parsing logic into a util module
